#12
modified filework with vehicles to more dynamically determine the classification of the new vehicle. Note, this only sets the variable to the classification of an object in the database whos brand and mode match the new vehicle. 

also updated autofill vehicles to be more efficient without running lines of code unnecessarily. 
created a new function is_complete for autofill_vehicle