#20 close #12 Progress
changed the txt files for #20
Made progress for #12
created a mower class and made tests for it.
tests need to be added to main_test.py
added the start to a decision tree for vehicle maker to create the correct vehicle object. needs more work
autofill needs another overhaul because at the moment it has no way to handle mower related numbers like deck sizes, hp ratings and that such.
might need to pass the classification along so that it has an idea what to look for.