#20 Close #12 Close #9 Close
for issue 20, the txt files have been renamed appropriately.
for issue #12 the build_vehicle function can now also build any type of ATV and Mower vehicle.
for issue #9 a decision making structure has been started that will allow for new types of vehicles to be created and specific information relating to those vehicles to be collected